Move IMAP sync into its own file
