+++ src/backends/imap/mod.rs
@@ -1,5 +1,7 @@
 mod imap;
+mod sync;
 mod watch;
 
 pub use imap::*;
+pub use sync::*;
 pub use watch::*;
